separate containers from components
