+++ src/containers/Details/__test__/PoetryDetails.test.js
