+++ src/containers/Details/style/PoetryDetails.css
